+++ src/feature/CardRegister/components/CardPreviewSection/CardPreviewContainer/CardPreviewContainer.tsx
@@ -3,20 +3,10 @@ import type {CardInfoType} from '../../../../../common/types/CardInfoType';
 import CardBrandLogo from '../CardBrandLogo/CardBrandLogo';
 import CardExpiryDateDisplay from '../CardExpiryDateDisplay/CardExpiryDateDisplay';
 import CardNumberDisplay from '../CardNumberDisplay/CardNumberDisplay';
+import {getBrandName} from '../../../domain/cardPolicy';
 
 const CardPreviewContainer = ({cardInfo}: {cardInfo: CardInfoType}) => {
   const {cardNumbers, expiryMonth, expiryYear} = cardInfo;
-
-  const getBrandName = (cardNumbers: [string, string, string, string]): 'visa' | 'masterCard' | null => {
-    const fullNumber = cardNumbers.map((chunk) => chunk.padEnd(4, '#')).join('');
-
-    if (fullNumber.startsWith('4')) return 'visa';
-
-    const prefix = Number(fullNumber.slice(0, 2));
-    if (prefix >= 51 && prefix <= 55) return 'masterCard';
-
-    return null;
-  };
 
   return (
     <Container>
